+++ HMI-APP/TouchGFX/gui/include/gui/dashboardscreen_screen/dashboardScreenPresenter.hpp
@@ -37,7 +37,7 @@ public:
   virtual void setIndicator(uint8_t index);
   virtual void setDriveMode(uint8_t index);
   virtual void setTripMode(uint8_t index);
-  virtual void setTripValue(uint32_t value);
+  virtual void setTripValue(uint16_t value);
   virtual void setReportMode(uint8_t index);
   virtual void setReportValue(uint16_t value);
