+++ HMI-APP/TouchGFX/gui/include/gui/model/ModelListener.hpp
@@ -25,7 +25,7 @@ public:
   virtual void setSignal(uint8_t percent) {};
   virtual void setDriveMode(uint8_t index) {};
   virtual void setTripMode(uint8_t index) {};
-  virtual void setTripValue(uint32_t value) {};
+  virtual void setTripValue(uint16_t value) {};
   virtual void setReportMode(uint8_t index) {};
   virtual void setReportValue(uint16_t value) {};
